requirements.txt
pyautogui
telebot
pyTelegramBotAPI
playsound==1.2.2
pillow
gtts
opencv-python
pipenv